# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SIAKAD GERBANG",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reference/organization-unit-histories": {
            "get": {
                "tags": [
                    "Reference/OrganizationUnitHistory"
                ],
                "summary": "List all OrganizationUnitHistory",
                "operationId": "722c3b1a3078915cc09de1c13baae0fb",
                "responses": {
                    "200": {
                        "description": "List of OrganizationUnitHistory",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationUnitHistoryListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reference/OrganizationUnitHistory"
                ],
                "summary": "Create a new OrganizationUnitHistory",
                "operationId": "dae511df217882798291e36b8f7c4b75",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "organization_unit_id",
                                    "organization_level_id",
                                    "start_date",
                                    "end_date",
                                    "is_active"
                                ],
                                "properties": {
                                    "organization_unit_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "organization_level_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "example": "example value"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "example": "example value"
                                    },
                                    "is_active": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "note": {
                                        "type": "object",
                                        "example": "example value"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OrganizationUnitHistory created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationUnitHistoryResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/reference/organization-unit-histories/{id}": {
            "get": {
                "tags": [
                    "Reference/OrganizationUnitHistory"
                ],
                "summary": "Get OrganizationUnitHistory detail by ID",
                "operationId": "3450c512a213b3ee33ecae5541eb9600",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrganizationUnitHistory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OrganizationUnitHistory detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationUnitHistoryResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "OrganizationUnitHistory not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reference/OrganizationUnitHistory"
                ],
                "summary": "Update an existing OrganizationUnitHistory",
                "operationId": "fb05000c30f0754d7de2dd06136149dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrganizationUnitHistory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "organization_unit_id",
                                    "organization_level_id",
                                    "start_date",
                                    "end_date",
                                    "is_active"
                                ],
                                "properties": {
                                    "organization_unit_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "organization_level_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "example": "example value"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "example": "example value"
                                    },
                                    "is_active": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "note": {
                                        "type": "object",
                                        "example": "example value"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OrganizationUnitHistory updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationUnitHistoryResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "OrganizationUnitHistory not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reference/OrganizationUnitHistory"
                ],
                "summary": "Delete a OrganizationUnitHistory by ID",
                "operationId": "4dd64689017acbfa8307d3b2e683a8cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrganizationUnitHistory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OrganizationUnitHistory deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Data deleted",
                                    "data": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "OrganizationUnitHistory not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrganizationLevel": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "properties": {
                            "en": {
                                "type": "string",
                                "example": "BLU"
                            },
                            "id": {
                                "type": "string",
                                "example": "Badan Layanan Umum"
                            }
                        },
                        "type": "object"
                    },
                    "type": {
                        "type": "string",
                        "example": "user"
                    }
                },
                "type": "object"
            },
...
            "OrganizationUnit": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "properties": {
                            "en": {
                                "type": "string",
                                "example": "Faculty Information"
                            },
                            "id": {
                                "type": "string",
                                "example": "Fakultas Informasi"
                            }
                        },
                        "type": "object"
                    },
                    "unit_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "unit_type": {
                        "$ref": "#/components/schemas/UnitType"
                    },
                    "parent": {
                        "$ref": "#/components/schemas/OrganizationUnit"
                    }
                },
                "type": "object"
            },
...
            "OrganizationUnitHistory": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "organization_unit_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "organization_level_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-14"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-14"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "note": {
                        "properties": {
                            "en": {
                                "type": "string",
                                "example": "Faculty Information"
                            },
                            "id": {
                                "type": "string",
                                "example": "Fakultas Informasi"
                            }
                        },
                        "type": "object"
                    },
                    "organization_level": {
                        "$ref": "#/components/schemas/OrganizationLevel"
                    },
                    "organization_unit": {
                        "$ref": "#/components/schemas/OrganizationUnit"
                    }
                },
                "type": "object"
            },
            "OrganizationUnitHistoryResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Organization unit history detail"
                    },
                    "data": {
                        "$ref": "#/components/schemas/OrganizationUnitHistory"
                    }
                },
                "type": "object"
            },
            "OrganizationUnitHistoryListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "List of organization unit histories"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrganizationUnitHistory"
                        }
                    }
                },
                "type": "object"
            },
...
            "UnitType": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Faculty"
                    },
                    "note": {
                        "type": "string",
                        "example": "Fakultas"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}